Download Agency Data

Download a single agency's data from the Missouri Attorney General's Office.

mv_get_agency(browser, agency, statistic, pause = 3, add_agency = TRUE)

Arguments

browser
Name of a browser binding created with RSelenium

agency
An agency name (for St. Louis City and County agencies) or id number (for agencies statewide).

statistic
The statistic to download

pause
Number of seconds to pause while pages load - adjust as needed depending on internet connection

add_agency
Append agency name to returned data; set to FALSE if you are downloading data outside of St. Louis City or County

Value

A tibble in 'long' format containing vehicle stop statistics for the requested agency. Data may need additional formatting - see mv_reformat.

Details

The statistics that can be downloaded for each agency are as follows:

"Disparity"

"Stops"

"Searches"

"Arrests"

"Contraband"